Reworked weapon pick ups
Hotline Miami's weapon pick ups feel delayed because it only picks up a weapon when you let go of Mouse2, as opposed to when you press it.
The reason the game does this is because by holding Mouse2 for a few seconds you can drop a weapon silently in the area you are standing in. A feature that has no use whatsoever and was removed from the sequel.

To make weapon pick ups feel better and keep this functionality so to stay true to the original game; I reworked the system so that you can silently drop a weapon by holding Alt + Mouse2 for a few seconds. If Alt is not being held before Mouse2 gets pressed the weapon will instantly be thrown the moment the game registers a press from Mouse2.