Revamped Biker's bossfight.
Fixed all of the issues with Biker's bossfights caused by the new shadow manager object. His shadows now behave as expected.
The Cleaver's collision with walls is also fixed. It now sticks to the walls properly and Biker's angle when prying it is also what it should be.
I removed the random chance the Cleaver has of bouncing off the wall because it's annoying and dumb.