Commented out references to "objGame" in other objects.
This will make the code less spaghetti by not making other objects dependent on its existence. Allowing for there to be project files that don't include it. It's also part of helping objGame not depend on Hotline Miami's structure so that it can potentially get used on other project as a generic debugging/development tool.